Remove duplicate code from mobile and desktop components
